# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GPSLink Gateway",
        "description": "Available API for GPSLink Gateway",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/current/create": {
            "post": {
                "tags": [
                    "Current"
                ],
                "summary": "Create Current Customer Info",
                "operationId": "CreateCurrentCustomer",
                "requestBody": {
                    "description": "Customer Customer Information",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "vehicle_assignment_id": {
                                        "type": "integer"
                                    },
                                    "customer_id": {
                                        "type": "integer"
                                    },
                                    "ipport_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "vehicle_assignment_id": 1,
                                    "customer_id": 1,
                                    "ipport_id": 1
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Current customer is successfully registered",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CurrentCustomer"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "409": {
                        "description": "Current customer already exist!"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/current/currentCustById/{id}": {
            "get": {
                "tags": [
                    "Current"
                ],
                "summary": "Get current customer by id",
                "operationId": "GetCurrentCustomerById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CurrentCustomer"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Current Customer not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/current/list": {
            "post": {
                "tags": [
                    "Current"
                ],
                "summary": "Get list of Currents",
                "operationId": "CurrentList",
                "requestBody": {
                    "description": "Customer Id - NOTE: If customer_id object is omitted then all current customers will be return.",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "customer_id": {
                                        "type": "integer"
                                    },
                                    "transporter_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "customer_id": 0,
                                    "transporter_id": 0
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/current/update/{id}": {
            "put": {
                "tags": [
                    "Current"
                ],
                "summary": "Updated Current",
                "description": "Update Current information.",
                "operationId": "UpdateCurrent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated Current Customer object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "vehicle_assignment_id": {
                                        "type": "integer"
                                    },
                                    "customer_id": {
                                        "type": "integer"
                                    },
                                    "ipport_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "id": 1,
                                    "vehicle_assignment_id": 1,
                                    "customer_id": 1,
                                    "ipport_id": 1
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Current customer is successfully updated!"
                    },
                    "404": {
                        "description": "Current customer not found"
                    },
                    "409": {
                        "description": "Current customer Id does not matched!"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/current/delete/{id}": {
            "delete": {
                "tags": [
                    "Current"
                ],
                "summary": "Delete Current by id",
                "operationId": "DeleteCurrent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Current customer is successfully deleted!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CurrentCustomer"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Current customer does not exist!"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "CurrentCustomer": {
                "title": "CurrentCustomer",
                "description": "CurrentCustomer",
                "required": [
                    "vehicle_assignment_id",
                    "customer_id",
                    "ipport_id"
                ],
                "properties": {
                    "id": {
                        "description": "ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "vehicle_assignment_id": {
                        "description": "Vehicle Assignment ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "customer_id": {
                        "description": "Customer ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "ipport_id": {
                        "description": "IP and Port ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "register_by_user_id": {
                        "description": "Register By UserId",
                        "type": "integer",
                        "format": "int64"
                    },
                    "updated_by_user_id": {
                        "description": "Updated By UserId",
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}